fix alm global configuration
