+++ src/main/java/com/microfocus/application/automation/tools/commonResultUpload/CommonResultUploadBuilder.java
@@ -36,7 +36,7 @@ import com.cloudbees.plugins.credentials.domains.URIRequirementBuilder;
 import com.cloudbees.plugins.credentials.matchers.IdMatcher;
 import com.microfocus.application.automation.tools.commonResultUpload.uploader.Uploader;
 import com.microfocus.application.automation.tools.model.AlmServerSettingsModel;
-import com.microfocus.application.automation.tools.settings.AlmServerSettingsBuilder;
+import com.microfocus.application.automation.tools.settings.AlmServerSettingsGlobalConfiguration;
 import hudson.Extension;
 import hudson.FilePath;
 import hudson.Launcher;
@@ -136,8 +136,7 @@ public class CommonResultUploadBuilder extends Recorder implements SimpleBuildSt
     }
 
     private String getAlmServerUrl(String almServerName) {
-        AlmServerSettingsModel[] almServers = Hudson.getInstance().getDescriptorByType(
-                AlmServerSettingsBuilder.DescriptorImpl.class).getInstallations();
+        AlmServerSettingsModel[] almServers = AlmServerSettingsGlobalConfiguration.getInstance().getInstallations();
         if (almServers != null && almServers.length > 0) {
             for (AlmServerSettingsModel almServerModel: almServers) {
                 if (almServerName.equalsIgnoreCase(almServerModel.getAlmServerName())) {
@@ -178,13 +177,11 @@ public class CommonResultUploadBuilder extends Recorder implements SimpleBuildSt
         }
 
         public boolean hasAlmServers() {
-            return Hudson.getInstance().getDescriptorByType(
-                    AlmServerSettingsBuilder.DescriptorImpl.class).hasAlmServers();
+            return AlmServerSettingsGlobalConfiguration.getInstance().hasAlmServers();
         }
 
         public AlmServerSettingsModel[] getAlmServers() {
-            return Hudson.getInstance().getDescriptorByType(
-                    AlmServerSettingsBuilder.DescriptorImpl.class).getInstallations();
+            return AlmServerSettingsGlobalConfiguration.getInstance().getInstallations();
         }
 
         public FormValidation doCheckAlmUserName(@QueryParameter String value) {
